Create packagees to separate data-types in db-folder.Change some property type from Int to Double. Delete unneeded properties in CurrentWeatherEntry.class
