+++ app/src/main/java/com/paolomanlunas/forecaster/data/WeatherApiService.kt
@@ -2,7 +2,7 @@ package com.paolomanlunas.forecaster.data
 
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
 import com.paolomanlunas.forecaster.BuildConfig
-import com.paolomanlunas.forecaster.data.response.CurrentWeatherResponse
+import com.paolomanlunas.forecaster.data.network.response.CurrentWeatherResponse
 import kotlinx.coroutines.Deferred
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
