--- app/src/main/java/com/paolomanlunas/forecaster/data/response/CurrentWeatherResponse.kt
@@ -1,11 +1,0 @@
-package com.paolomanlunas.forecaster.data.response
-
-import com.google.gson.annotations.SerializedName
-
-
-data class CurrentWeatherResponse(
-    @SerializedName("current")
-    val currentWeatherEntry: CurrentWeatherEntry,
-    val location: Location,
-    val request: Request
-)
